backend/pyproject.toml
[project]
name = "recipe-ai-backend"
version = "0.1.0"
description = "AI-powered recipe generation backend"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "slowapi>=0.1.9",
    "firebase-admin>=7.1.0",
    "google-cloud-storage>=3.8.0",
    "litellm>=1.81.0",
    "pydantic>=2.12.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.0",
    "cachetools>=6.2.4",
    "google-genai>=1.60.0",
]


[dependency-groups]
dev = [
    "ipython>=9.0.0",
    "jupyter>=1.1.0",
]
